Use feed library for Hive Atom feed
