+++ package.json
@@ -27,6 +27,7 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "d3": "^7.9.0",
+    "feed": "^5.1.0",
     "lucide-react": "^0.542.0",
     "next": "^15.5.0",
     "papaparse": "^5.5.3",
